Refactor da queue tests
