+++ bin/citrea/tests/bitcoin/mod.rs
@@ -13,6 +13,7 @@ pub mod fork;
 #[cfg(feature = "testing")]
 pub mod full_node;
 pub mod guest_cycles;
+pub mod queue;
 pub mod sequencer_commitments;
 pub mod sequencer_test;
 pub mod syncing;
